@@ -14,7 +14,7 @@ class Raven_SerializerTestObject
     private $foo = 'bar';
 }
 
-class Raven_Tests_SerializerTest extends PHPUnit_Framework_TestCase
+class Raven_Tests_SerializerTest extends \PHPUnit\Framework\TestCase
 {
     public function testArraysAreArrays()
     {
@@ -24,12 +24,21 @@ class Raven_Tests_SerializerTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(array('1', '2', '3'), $result);
     }
 
+    public function testStdClassAreArrays()
+    {
+        $serializer = new Raven_Serializer();
+        $input = new stdClass();
+        $input->foo = 'BAR';
+        $result = $serializer->serialize($input);
+        $this->assertEquals(array('foo' => 'BAR'), $result);
+    }
+
     public function testObjectsAreStrings()
     {
         $serializer = new Raven_Serializer();
-        $input = new Raven_StacktraceTestObject();
+        $input = new Raven_SerializerTestObject();
         $result = $serializer->serialize($input);
-        $this->assertEquals('Object Raven_StacktraceTestObject', $result);
+        $this->assertEquals('Object Raven_SerializerTestObject', $result);
     }
 
     public function testIntsAreInts()
@@ -37,7 +46,7 @@ class Raven_Tests_SerializerTest extends PHPUnit_Framework_TestCase
         $serializer = new Raven_Serializer();
         $input = 1;
         $result = $serializer->serialize($input);
-        $this->assertTrue(is_integer($result));
+        $this->assertInternalType('integer', $result);
         $this->assertEquals(1, $result);
     }
 
@@ -46,7 +55,7 @@ class Raven_Tests_SerializerTest extends PHPUnit_Framework_TestCase
         $serializer = new Raven_Serializer();
         $input = 1.5;
         $result = $serializer->serialize($input);
-        $this->assertTrue(is_float($result));
+        $this->assertInternalType('double', $result);
         $this->assertEquals(1.5, $result);
     }
 
@@ -55,13 +64,11 @@ class Raven_Tests_SerializerTest extends PHPUnit_Framework_TestCase
         $serializer = new Raven_Serializer();
         $input = true;
         $result = $serializer->serialize($input);
-        $this->assertTrue(is_bool($result));
-        $this->assertEquals(true, $result);
+        $this->assertTrue($result);
 
         $input = false;
         $result = $serializer->serialize($input);
-        $this->assertTrue(is_bool($result));
-        $this->assertEquals(false, $result);
+        $this->assertFalse($result);
     }
 
     public function testNull()
@@ -69,8 +76,7 @@ class Raven_Tests_SerializerTest extends PHPUnit_Framework_TestCase
         $serializer = new Raven_Serializer();
         $input = null;
         $result = $serializer->serialize($input);
-        $this->assertTrue(is_null($result));
-        $this->assertEquals(null, $result);
+        $this->assertNull($result);
     }
 
     public function testRecursionMaxDepth()
@@ -80,5 +86,119 @@ class Raven_Tests_SerializerTest extends PHPUnit_Framework_TestCase
         $input[] = &$input;
         $result = $serializer->serialize($input, 3);
         $this->assertEquals(array(array(array('Array of length 1'))), $result);
+    }
+
+    public function testObjectInArray()
+    {
+        $serializer = new Raven_Serializer();
+        $input = array('foo' => new Raven_Serializer());
+        $result = $serializer->serialize($input);
+        $this->assertEquals(array('foo' => 'Object Raven_Serializer'), $result);
+    }
+
+    /**
+     * @covers Raven_Serializer::serializeString
+     */
+    public function testBrokenEncoding()
+    {
+        $serializer = new Raven_Serializer();
+        foreach (array('7efbce4384', 'b782b5d8e5', '9dde8d1427', '8fd4c373ca', '9b8e84cb90') as $key) {
+            $input = pack('H*', $key);
+            $result = $serializer->serialize($input);
+            $this->assertInternalType('string', $result);
+            if (function_exists('mb_detect_encoding')) {
+                $this->assertContains(mb_detect_encoding($result), array('ASCII', 'UTF-8'));
+            }
+        }
+    }
+
+    /**
+     * @covers Raven_Serializer::serializeString
+     */
+    public function testLongString()
+    {
+        $serializer = new Raven_Serializer();
+        for ($i = 0; $i < 100; $i++) {
+            foreach (array(100, 1000, 1010, 1024, 1050, 1100, 10000) as $length) {
+                $input = '';
+                for ($j = 0; $j < $length; $j++) {
+                    $input .= chr(mt_rand(ord('a'), ord('z')));
+                }
+                $result = $serializer->serialize($input);
+                $this->assertInternalType('string', $result);
+                $this->assertLessThanOrEqual(1024, strlen($result));
+            }
+        }
+    }
+
+    /**
+     * @covers Raven_Serializer::serializeString
+     */
+    public function testLongStringWithOverwrittenMessageLength()
+    {
+        $serializer = new Raven_Serializer();
+        $serializer->setMessageLimit(500);
+        for ($i = 0; $i < 100; $i++) {
+            foreach (array(100, 490, 499, 500, 501, 1000, 10000) as $length) {
+                $input = '';
+                for ($j = 0; $j < $length; $j++) {
+                    $input .= chr(mt_rand(ord('a'), ord('z')));
+                }
+                $result = $serializer->serialize($input);
+                $this->assertInternalType('string', $result);
+                $this->assertLessThanOrEqual(500, strlen($result));
+            }
+        }
+    }
+
+    /**
+     * @covers Raven_Serializer::serializeString
+     */
+    public function testLongStringWithOverwrittenMessageLengthZero()
+    {
+        $serializer = new Raven_Serializer();
+        $serializer->setMessageLimit(0);
+        for ($i = 0; $i < 100; $i++) {
+            foreach (array(100, 490, 499, 500, 501, 1000, 10000) as $length) {
+                $input = '';
+                for ($j = 0; $j < $length; $j++) {
+                    $input .= chr(mt_rand(ord('a'), ord('z')));
+                }
+                $result = $serializer->serialize($input);
+                $this->assertInternalType('string', $result);
+                $this->assertEquals($length, strlen($result));
+            }
+        }
+    }
+
+    /**
+     * @covers Raven_Serializer::serializeValue
+     */
+    public function testSerializeValueResource()
+    {
+        $serializer = new Raven_Serializer();
+        $filename = tempnam(sys_get_temp_dir(), 'sentry_test_');
+        $fo = fopen($filename, 'wb');
+
+        $result = $serializer->serialize($fo);
+        $this->assertInternalType('string', $result);
+        $this->assertEquals('Resource stream', $result);
+    }
+
+    public function testClippingUTF8Characters()
+    {
+        if (!extension_loaded('mbstring')) {
+            $this->markTestSkipped('mbstring extension is not enabled.');
+        }
+
+        $teststring = 'Прекратите надеяться, что ваши пользователи будут сообщать об ошибках';
+        $serializer = new Raven_Serializer(null, 19); // Length of 19 will clip character in half if no mb_* string functions are used for the teststring
+
+        $clipped = $serializer->serialize($teststring);
+        $this->assertEquals('Прекратит {clipped}', $clipped);
+
+        Raven_Compat::json_encode($clipped);
+
+        $this->assertEquals(JSON_ERROR_NONE, json_last_error());
     }
 }
