@@ -21,7 +21,6 @@ class Raven_Autoloader
      */
     public static function register()
     {
-        ini_set('unserialize_callback_func', 'spl_autoload_call');
         spl_autoload_register(array('Raven_Autoloader', 'autoload'));
     }
 
@@ -32,11 +31,13 @@ class Raven_Autoloader
      */
     public static function autoload($class)
     {
-        if (0 !== strpos($class, 'Raven')) {
+        if (substr($class, 0, 6) !== 'Raven_') {
             return;
         }
 
-        if (is_file($file = dirname(__FILE__).'/../'.str_replace(array('_', "\0"), array('/', ''), $class).'.php')) {
+        $file = dirname(__FILE__).'/../'.str_replace(array('_', "\0"), array('/', ''), $class).'.php';
+        if (is_file($file)) {
+            /** @noinspection PhpIncludeInspection */
             require $file;
         }
     }
