@@ -16,7 +16,8 @@
 
 class Raven_Client
 {
-    const VERSION = '0.14.0.dev0';
+    const VERSION = '1.10.x-dev';
+
     const PROTOCOL = '6';
 
     const DEBUG = 'debug';
@@ -28,26 +29,131 @@ class Raven_Client
 
     const MESSAGE_LIMIT = 1024;
 
-    public $severity_map;
+    public $breadcrumbs;
+    /**
+     * @var Raven_Context
+     */
+    public $context;
     public $extra_data;
-
+    /**
+     * @var array|null
+     */
+    public $severity_map;
     public $store_errors_for_bulk_send = false;
 
-    public function __construct($options_or_dsn=null, $options=array())
+    protected $error_handler;
+    protected $error_types;
+
+    /**
+     * @var Raven_Serializer
+     */
+    protected $serializer;
+    /**
+     * @var Raven_ReprSerializer
+     */
+    protected $reprSerializer;
+
+    /**
+     * @var string
+     */
+    protected $app_path;
+    /**
+     * @var string[]
+     */
+    protected $prefixes;
+    /**
+     * @var string[]|null
+     */
+    protected $excluded_app_paths;
+    /**
+     * @var Callable
+     */
+    protected $transport;
+
+    public $logger;
+    /**
+     * @var string Full URL to Sentry
+     */
+    public $server;
+    public $secret_key;
+    public $public_key;
+    public $project;
+    public $auto_log_stacks;
+    public $name;
+    public $site;
+    public $tags;
+    public $release;
+    public $environment;
+    public $sample_rate;
+    public $trace;
+    public $timeout;
+    public $message_limit;
+    public $exclude;
+    public $excluded_exceptions;
+    public $http_proxy;
+    public $ignore_server_port;
+    protected $send_callback;
+    public $curl_method;
+    public $curl_path;
+    public $curl_ipv4;
+    public $ca_cert;
+    public $verify_ssl;
+    public $curl_ssl_version;
+    public $trust_x_forwarded_proto;
+    public $mb_detect_order;
+    /**
+     * @var Raven_Processor[]
+     */
+    public $processors;
+    /**
+     * @var string|int|null
+     */
+    public $_lasterror;
+    /**
+     * @var object|null
+     */
+    protected $_last_sentry_error;
+    public $_last_event_id;
+    public $_user;
+    public $_pending_events;
+    public $sdk;
+    /**
+     * @var Raven_CurlHandler
+     */
+    protected $_curl_handler;
+    /**
+     * @var resource|null
+     */
+    protected $_curl_instance;
+    /**
+     * @var bool
+     */
+    protected $_shutdown_function_has_been_set;
+
+    /**
+     * @var bool
+     */
+    public $useCompression;
+
+    public function __construct($options_or_dsn = null, $options = array())
     {
-        if (is_null($options_or_dsn) && !empty($_SERVER['SENTRY_DSN'])) {
-            // Read from environment
-            $options_or_dsn = $_SERVER['SENTRY_DSN'];
+        if (is_array($options_or_dsn)) {
+            $options = array_merge($options_or_dsn, $options);
         }
-        if (!is_array($options_or_dsn)) {
-            if (!empty($options_or_dsn)) {
-                // Must be a valid DSN
-                $options_or_dsn = self::parseDSN($options_or_dsn);
-            } else {
-                $options_or_dsn = array();
-            }
+
+        if (!is_array($options_or_dsn) && !empty($options_or_dsn)) {
+            $dsn = $options_or_dsn;
+        } elseif (!empty($_SERVER['SENTRY_DSN'])) {
+            $dsn = @$_SERVER['SENTRY_DSN'];
+        } elseif (!empty($options['dsn'])) {
+            $dsn = $options['dsn'];
+        } else {
+            $dsn = null;
         }
-        $options = array_merge($options_or_dsn, $options);
+
+        if (!empty($dsn)) {
+            $options = array_merge($options, self::parseDSN($dsn));
+        }
 
         $this->logger = Raven_Util::get($options, 'logger', 'php');
         $this->server = Raven_Util::get($options, 'server');
@@ -56,42 +162,275 @@ class Raven_Client
         $this->project = Raven_Util::get($options, 'project', 1);
         $this->auto_log_stacks = (bool) Raven_Util::get($options, 'auto_log_stacks', false);
         $this->name = Raven_Util::get($options, 'name', Raven_Compat::gethostname());
-        $this->site = Raven_Util::get($options, 'site', $this->_server_variable('SERVER_NAME'));
+        $this->site = Raven_Util::get($options, 'site', self::_server_variable('SERVER_NAME'));
         $this->tags = Raven_Util::get($options, 'tags', array());
-        $this->release = Raven_util::get($options, 'release', null);
-        $this->environment = Raven_util::get($options, 'environment', null);
+        $this->release = Raven_Util::get($options, 'release', null);
+        $this->environment = Raven_Util::get($options, 'environment', null);
+        $this->sample_rate = Raven_Util::get($options, 'sample_rate', 1);
         $this->trace = (bool) Raven_Util::get($options, 'trace', true);
         $this->timeout = Raven_Util::get($options, 'timeout', 2);
         $this->message_limit = Raven_Util::get($options, 'message_limit', self::MESSAGE_LIMIT);
         $this->exclude = Raven_Util::get($options, 'exclude', array());
+        $this->excluded_exceptions = Raven_Util::get($options, 'excluded_exceptions', array());
         $this->severity_map = null;
-        $this->shift_vars = (bool) Raven_Util::get($options, 'shift_vars', true);
         $this->http_proxy = Raven_Util::get($options, 'http_proxy');
+        $this->ignore_server_port = Raven_Util::get($options, 'ignore_server_port', false);
         $this->extra_data = Raven_Util::get($options, 'extra', array());
         $this->send_callback = Raven_Util::get($options, 'send_callback', null);
         $this->curl_method = Raven_Util::get($options, 'curl_method', 'sync');
         $this->curl_path = Raven_Util::get($options, 'curl_path', 'curl');
-        $this->curl_ipv4 = Raven_util::get($options, 'curl_ipv4', true);
-        $this->ca_cert = Raven_util::get($options, 'ca_cert', $this->get_default_ca_cert());
-        $this->verify_ssl = Raven_util::get($options, 'verify_ssl', true);
+        $this->curl_ipv4 = Raven_Util::get($options, 'curl_ipv4', true);
+        $this->ca_cert = Raven_Util::get($options, 'ca_cert', static::get_default_ca_cert());
+        $this->verify_ssl = Raven_Util::get($options, 'verify_ssl', true);
         $this->curl_ssl_version = Raven_Util::get($options, 'curl_ssl_version');
         $this->trust_x_forwarded_proto = Raven_Util::get($options, 'trust_x_forwarded_proto');
+        $this->transport = Raven_Util::get($options, 'transport', null);
+        $this->mb_detect_order = Raven_Util::get($options, 'mb_detect_order', null);
+        $this->error_types = Raven_Util::get($options, 'error_types', null);
 
+        // app path is used to determine if code is part of your application
+        $this->setAppPath(Raven_Util::get($options, 'app_path', null));
+        $this->setExcludedAppPaths(Raven_Util::get($options, 'excluded_app_paths', null));
+        // a list of prefixes used to coerce absolute paths into relative
+        $this->setPrefixes(Raven_Util::get($options, 'prefixes', static::getDefaultPrefixes()));
         $this->processors = $this->setProcessorsFromOptions($options);
 
         $this->_lasterror = null;
+        $this->_last_sentry_error = null;
+        $this->_curl_instance = null;
+        $this->_last_event_id = null;
         $this->_user = null;
+        $this->_pending_events = array();
         $this->context = new Raven_Context();
+        $this->breadcrumbs = new Raven_Breadcrumbs();
+        $this->_shutdown_function_has_been_set = false;
+        $this->useCompression = function_exists('gzcompress');
+
+        $this->sdk = Raven_Util::get($options, 'sdk', array(
+            'name' => 'sentry-php',
+            'version' => self::VERSION,
+        ));
+        $this->serializer = new Raven_Serializer($this->mb_detect_order, $this->message_limit);
+        $this->reprSerializer = new Raven_ReprSerializer($this->mb_detect_order, $this->message_limit);
 
         if ($this->curl_method == 'async') {
             $this->_curl_handler = new Raven_CurlHandler($this->get_curl_options());
         }
+
+        $this->transaction = new Raven_TransactionStack();
+        if (static::is_http_request() && isset($_SERVER['PATH_INFO'])) {
+            // @codeCoverageIgnoreStart
+            $this->transaction->push($_SERVER['PATH_INFO']);
+            // @codeCoverageIgnoreEnd
+        }
+
+        if (Raven_Util::get($options, 'install_default_breadcrumb_handlers', true)) {
+            $this->registerDefaultBreadcrumbHandlers();
+        }
+
+        if (Raven_Util::get($options, 'install_shutdown_handler', true)) {
+            $this->registerShutdownFunction();
+        }
+
+        $this->triggerAutoload();
     }
 
+    public function __destruct()
+    {
+        // Force close curl resource
+        $this->close_curl_resource();
+    }
+
+    /**
+     * Destruct all objects contain link to this object
+     *
+     * This method can not delete shutdown handler
+     */
+    public function close_all_children_link()
+    {
+        $this->processors = array();
+    }
+
+    /**
+     * Installs any available automated hooks (such as error_reporting).
+     */
+    public function install()
+    {
+        if ($this->error_handler) {
+            throw new Raven_Exception(sprintf('%s->install() must only be called once', get_class($this)));
+        }
+        $this->error_handler = new Raven_ErrorHandler($this, false, $this->error_types);
+        $this->error_handler->registerExceptionHandler();
+        $this->error_handler->registerErrorHandler();
+        $this->error_handler->registerShutdownFunction();
+
+        if ($this->_curl_handler) {
+            $this->_curl_handler->registerShutdownFunction();
+        }
+
+        return $this;
+    }
+
+    public function getRelease()
+    {
+        return $this->release;
+    }
+
+    public function setRelease($value)
+    {
+        $this->release = $value;
+        return $this;
+    }
+
+    public function getEnvironment()
+    {
+        return $this->environment;
+    }
+
+    public function setEnvironment($value)
+    {
+        $this->environment = $value;
+        return $this;
+    }
+
+    /**
+     * Note: Prior to PHP 5.6, a stream opened with php://input can
+     * only be read once;
+     *
+     * @see http://php.net/manual/en/wrappers.php.php
+     */
+    protected static function getInputStream()
+    {
+        if (PHP_VERSION_ID < 50600) {
+            return null;
+        }
+
+        return file_get_contents('php://input');
+    }
+
+    private static function getDefaultPrefixes()
+    {
+        $value = get_include_path();
+        return explode(PATH_SEPARATOR, $value);
+    }
+
+    private static function _convertPath($value)
+    {
+        $path = @realpath($value);
+        if ($path === false) {
+            $path = $value;
+        }
+        // we need app_path to have a trailing slash otherwise
+        // base path detection becomes complex if the same
+        // prefix is matched
+        if ($path{0} === DIRECTORY_SEPARATOR && substr($path, -1) !== DIRECTORY_SEPARATOR) {
+            $path .= DIRECTORY_SEPARATOR;
+        }
+        return $path;
+    }
+
+    public function getAppPath()
+    {
+        return $this->app_path;
+    }
+
+    public function setAppPath($value)
+    {
+        if ($value) {
+            $this->app_path = static::_convertPath($value);
+        } else {
+            $this->app_path = null;
+        }
+        return $this;
+    }
+
+    public function getExcludedAppPaths()
+    {
+        return $this->excluded_app_paths;
+    }
+
+    public function setExcludedAppPaths($value)
+    {
+        if ($value) {
+            $excluded_app_paths = array();
+
+            // We should be able to exclude a php files
+            foreach ((array) $value as $path) {
+                $excluded_app_paths[] = substr($path, -4) !== '.php' ? self::_convertPath($path) : $path;
+            }
+        } else {
+            $excluded_app_paths = null;
+        }
+
+        $this->excluded_app_paths = $excluded_app_paths;
+
+        return $this;
+    }
+
+    public function getPrefixes()
+    {
+        return $this->prefixes;
+    }
+
+    /**
+     * @param array $value
+     * @return Raven_Client
+     */
+    public function setPrefixes($value)
+    {
+        $this->prefixes = $value ? array_map(array($this, '_convertPath'), $value) : $value;
+        return $this;
+    }
+
+    public function getSendCallback()
+    {
+        return $this->send_callback;
+    }
+
+    public function setSendCallback($value)
+    {
+        $this->send_callback = $value;
+        return $this;
+    }
+
+    public function getTransport()
+    {
+        return $this->transport;
+    }
+
+    public function getServerEndpoint($value = '')
+    {
+        return $this->server;
+    }
+
+    public static function getUserAgent()
+    {
+        return 'sentry-php/' . self::VERSION;
+    }
+
+    /**
+     * Set a custom transport to override how Sentry events are sent upstream.
+     *
+     * The bound function will be called with ``$client`` and ``$data`` arguments
+     * and is responsible for encoding the data, authenticating, and sending
+     * the data to the upstream Sentry server.
+     *
+     * @param Callable $value Function to be called
+     * @return Raven_Client
+     */
+    public function setTransport($value)
+    {
+        $this->transport = $value;
+        return $this;
+    }
+
+    /**
+     * @return string[]|Raven_Processor[]
+     */
     public static function getDefaultProcessors()
     {
         return array(
-            'Raven_SanitizeDataProcessor',
+            'Raven_Processor_SanitizeDataProcessor',
         );
     }
 
@@ -100,16 +439,22 @@ class Raven_Client
      * sent to Sentry.
      *
      * @param $options
-     * @return array
+     * @return Raven_Processor[]
      */
     public function setProcessorsFromOptions($options)
     {
         $processors = array();
-        foreach (Raven_util::get($options, 'processors', self::getDefaultProcessors()) as $processor) {
+        foreach (Raven_Util::get($options, 'processors', static::getDefaultProcessors()) as $processor) {
+            /**
+             * @var Raven_Processor        $new_processor
+             * @var Raven_Processor|string $processor
+             */
             $new_processor = new $processor($this);
 
             if (isset($options['processorOptions']) && is_array($options['processorOptions'])) {
-                if (isset($options['processorOptions'][$processor]) && method_exists($processor, 'setProcessorOptions')) {
+                if (isset($options['processorOptions'][$processor])
+                    && method_exists($processor, 'setProcessorOptions')
+                ) {
                     $new_processor->setProcessorOptions($options['processorOptions'][$processor]);
                 }
             }
@@ -121,18 +466,34 @@ class Raven_Client
     /**
      * Parses a Raven-compatible DSN and returns an array of its values.
      *
-     * @param string    $dsn    Raven compatible DSN: http://raven.readthedocs.org/en/latest/config/#the-sentry-dsn
-     * @return array            parsed DSN
+     * @param string $dsn Raven compatible DSN
+     * @return array      parsed DSN
+     *
+     * @see http://raven.readthedocs.org/en/latest/config/#the-sentry-dsn
      */
     public static function parseDSN($dsn)
     {
+        switch (strtolower($dsn)) {
+            case '':
+            case 'false':
+            case '(false)':
+            case 'empty':
+            case '(empty)':
+            case 'null':
+            case '(null)':
+                return array();
+        }
+
         $url = parse_url($dsn);
         $scheme = (isset($url['scheme']) ? $url['scheme'] : '');
         if (!in_array($scheme, array('http', 'https'))) {
-            throw new InvalidArgumentException('Unsupported Sentry DSN scheme: ' . (!empty($scheme) ? $scheme : '<not set>'));
+            throw new InvalidArgumentException(
+                'Unsupported Sentry DSN scheme: '.
+                (!empty($scheme) ? $scheme : '<not set>')
+            );
         }
         $netloc = (isset($url['host']) ? $url['host'] : null);
-        $netloc.= (isset($url['port']) ? ':'.$url['port'] : null);
+        $netloc .= (isset($url['port']) ? ':'.$url['port'] : null);
         $rawpath = (isset($url['path']) ? $url['path'] : null);
         if ($rawpath) {
             $pos = strrpos($rawpath, '/', 1);
@@ -149,7 +510,7 @@ class Raven_Client
         }
         $username = (isset($url['user']) ? $url['user'] : null);
         $password = (isset($url['pass']) ? $url['pass'] : null);
-        if (empty($netloc) || empty($project) || empty($username) || empty($password)) {
+        if (empty($netloc) || empty($project) || empty($username)) {
             throw new InvalidArgumentException('Invalid Sentry DSN: ' . $dsn);
         }
 
@@ -168,6 +529,10 @@ class Raven_Client
 
     /**
      * Given an identifier, returns a Sentry searchable string.
+     *
+     * @param mixed $ident
+     * @return mixed
+     * @codeCoverageIgnore
      */
     public function getIdent($ident)
     {
@@ -176,16 +541,26 @@ class Raven_Client
     }
 
     /**
-     * Deprecated
+     * @param string     $message The message (primary description) for the event.
+     * @param array      $params  params to use when formatting the message.
+     * @param string     $level   Log level group
+     * @param bool|array $stack
+     * @param mixed      $vars
+     * @return string|null
+     * @deprecated
+     * @codeCoverageIgnore
      */
-    public function message($message, $params=array(), $level=self::INFO,
-                            $stack=false, $vars = null)
+    public function message($message, $params = array(), $level = self::INFO,
+                            $stack = false, $vars = null)
     {
         return $this->captureMessage($message, $params, $level, $stack, $vars);
     }
 
     /**
-     * Deprecated
+     * @param Exception $exception
+     * @return string|null
+     * @deprecated
+     * @codeCoverageIgnore
      */
     public function exception($exception)
     {
@@ -194,9 +569,16 @@ class Raven_Client
 
     /**
      * Log a message to sentry
+     *
+     * @param string     $message The message (primary description) for the event.
+     * @param array      $params  params to use when formatting the message.
+     * @param array      $data    Additional attributes to pass with this event (see Sentry docs).
+     * @param bool|array $stack
+     * @param mixed      $vars
+     * @return string|null
      */
-    public function captureMessage($message, $params=array(), $level_or_options=array(),
-                            $stack=false, $vars = null)
+    public function captureMessage($message, $params = array(), $data = array(),
+                            $stack = false, $vars = null)
     {
         // Gracefully handle messages which contain formatting characters, but were not
         // intended to be used with formatting.
@@ -206,20 +588,20 @@ class Raven_Client
             $formatted_message = $message;
         }
 
-        if ($level_or_options === null) {
+        if ($data === null) {
             $data = array();
-        } elseif (!is_array($level_or_options)) {
+        // support legacy method of passing in a level name as the third arg
+        } elseif (!is_array($data)) {
             $data = array(
-                'level' => $level_or_options,
+                'level' => $data,
             );
-        } else {
-            $data = $level_or_options;
         }
 
         $data['message'] = $formatted_message;
         $data['sentry.interfaces.Message'] = array(
             'message' => $message,
             'params' => $params,
+            'formatted' => $formatted_message,
         );
 
         return $this->capture($data, $stack, $vars);
@@ -227,36 +609,36 @@ class Raven_Client
 
     /**
      * Log an exception to sentry
+     *
+     * @param \Throwable|\Exception $exception The Throwable/Exception object.
+     * @param array                 $data      Additional attributes to pass with this event (see Sentry docs).
+     * @param mixed                 $logger
+     * @param mixed                 $vars
+     * @return string|null
      */
-    public function captureException($exception, $culprit_or_options=null, $logger=null, $vars=null)
+    public function captureException($exception, $data = null, $logger = null, $vars = null)
     {
-        $has_chained_exceptions = version_compare(PHP_VERSION, '5.3.0', '>=');
+        $has_chained_exceptions = PHP_VERSION_ID >= 50300;
 
         if (in_array(get_class($exception), $this->exclude)) {
             return null;
         }
 
-        if (!is_array($culprit_or_options)) {
-            $data = array();
-            if ($culprit_or_options !== null) {
-                $data['culprit'] = $culprit_or_options;
+        foreach ($this->excluded_exceptions as $exclude) {
+            if ($exception instanceof $exclude) {
+                return null;
             }
-        } else {
-            $data = $culprit_or_options;
         }
 
-        // TODO(dcramer): DRY this up
-        $message = $exception->getMessage();
-        if (empty($message)) {
-            $message = get_class($exception);
+        if ($data === null) {
+            $data = array();
         }
 
         $exc = $exception;
         do {
             $exc_data = array(
-                'value' => $exc->getMessage(),
+                'value' => $this->serializer->serialize($exc->getMessage()),
                 'type' => get_class($exc),
-                'module' => $exc->getFile() .':'. $exc->getLine(),
             );
 
             /**'exception'
@@ -271,21 +653,16 @@ class Raven_Client
 
             array_unshift($trace, $frame_where_exception_thrown);
 
-            // manually trigger autoloading, as it's not done in some edge cases due to PHP bugs (see #60149)
-            if (!class_exists('Raven_Stacktrace')) {
-                spl_autoload_call('Raven_Stacktrace');
-            }
-
             $exc_data['stacktrace'] = array(
                 'frames' => Raven_Stacktrace::get_stack_info(
-                    $trace, $this->trace, $this->shift_vars, $vars, $this->message_limit
+                    $trace, $this->trace, $vars, $this->message_limit, $this->prefixes,
+                    $this->app_path, $this->excluded_app_paths, $this->serializer, $this->reprSerializer
                 ),
             );
 
             $exceptions[] = $exc_data;
         } while ($has_chained_exceptions && $exc = $exc->getPrevious());
 
-        $data['message'] = $message;
         $data['exception'] = array(
             'values' => array_reverse($exceptions),
         );
@@ -304,10 +681,33 @@ class Raven_Client
         return $this->capture($data, $trace, $vars);
     }
 
+
+    /**
+     * Capture the most recent error (obtained with ``error_get_last``).
+     * @return string|null
+     */
+    public function captureLastError()
+    {
+        if (null === $error = error_get_last()) {
+            return null;
+        }
+
+        $e = new ErrorException(
+            @$error['message'], 0, @$error['type'],
+            @$error['file'], @$error['line']
+        );
+
+        return $this->captureException($e);
+    }
+
     /**
      * Log an query to sentry
+     *
+     * @param string|null $query
+     * @param string      $level
+     * @param string      $engine
      */
-    public function captureQuery($query, $level=self::INFO, $engine = '')
+    public function captureQuery($query, $level = self::INFO, $engine = '')
     {
         $data = array(
             'message' => $query,
@@ -323,44 +723,73 @@ class Raven_Client
         return $this->capture($data, false);
     }
 
-    protected function is_http_request()
+    /**
+     * Return the last captured event's ID or null if none available.
+     */
+    public function getLastEventID()
+    {
+        return $this->_last_event_id;
+    }
+
+    protected function registerDefaultBreadcrumbHandlers()
+    {
+        $handler = new Raven_Breadcrumbs_ErrorHandler($this);
+        $handler->install();
+    }
+
+    protected function registerShutdownFunction()
+    {
+        if (!$this->_shutdown_function_has_been_set) {
+            $this->_shutdown_function_has_been_set = true;
+            register_shutdown_function(array($this, 'onShutdown'));
+        }
+    }
+
+    /**
+     * @return bool
+     * @codeCoverageIgnore
+     */
+    protected static function is_http_request()
     {
         return isset($_SERVER['REQUEST_METHOD']) && PHP_SAPI !== 'cli';
     }
 
     protected function get_http_data()
     {
-        $env = $headers = array();
+        $headers = array();
 
         foreach ($_SERVER as $key => $value) {
             if (0 === strpos($key, 'HTTP_')) {
-                $headers[str_replace(' ', '-', ucwords(strtolower(str_replace('_', ' ', substr($key, 5)))))] = $value;
+                $header_key =
+                    str_replace(' ', '-', ucwords(strtolower(str_replace('_', ' ', substr($key, 5)))));
+                $headers[$header_key] = $value;
             } elseif (in_array($key, array('CONTENT_TYPE', 'CONTENT_LENGTH')) && $value !== '') {
-                $headers[str_replace(' ', '-', ucwords(strtolower(str_replace('_', ' ', $key))))] = $value;
-            } else {
-                $env[$key] = $value;
+                $header_key = str_replace(' ', '-', ucwords(strtolower(str_replace('_', ' ', $key))));
+                $headers[$header_key] = $value;
             }
         }
 
         $result = array(
-            'method' => $this->_server_variable('REQUEST_METHOD'),
+            'method' => self::_server_variable('REQUEST_METHOD'),
             'url' => $this->get_current_url(),
-            'query_string' => $this->_server_variable('QUERY_STRING'),
+            'query_string' => self::_server_variable('QUERY_STRING'),
         );
 
         // dont set this as an empty array as PHP will treat it as a numeric array
         // instead of a mapping which goes against the defined Sentry spec
         if (!empty($_POST)) {
             $result['data'] = $_POST;
+        } elseif (isset($_SERVER['CONTENT_TYPE']) && stripos($_SERVER['CONTENT_TYPE'], 'application/json') === 0) {
+            $raw_data = $this->getInputStream() ?: false;
+            if ($raw_data !== false) {
+                $result['data'] = (array) json_decode($raw_data, true) ?: null;
+            }
         }
         if (!empty($_COOKIE)) {
             $result['cookies'] = $_COOKIE;
         }
         if (!empty($headers)) {
             $result['headers'] = $headers;
-        }
-        if (!empty($env)) {
-            $result['env'] = $env;
         }
 
         return array(
@@ -372,12 +801,15 @@ class Raven_Client
     {
         $user = $this->context->user;
         if ($user === null) {
-            if (!session_id()) {
+            if (!function_exists('session_id') || !session_id()) {
                 return array();
             }
             $user = array(
                 'id' => session_id(),
             );
+            if (!empty($_SERVER['REMOTE_ADDR'])) {
+                $user['ip_address'] = $_SERVER['REMOTE_ADDR'];
+            }
             if (!empty($_SESSION)) {
                 $user['data'] = $_SESSION;
             }
@@ -401,10 +833,12 @@ class Raven_Client
             'logger' => $this->logger,
             'tags' => $this->tags,
             'platform' => 'php',
+            'sdk' => $this->sdk,
+            'transaction' => $this->transaction->peek(),
         );
     }
 
-    public function capture($data, $stack, $vars = null)
+    public function capture($data, $stack = null, $vars = null)
     {
         if (!isset($data['timestamp'])) {
             $data['timestamp'] = gmdate('Y-m-d\TH:i:s\Z');
@@ -419,7 +853,7 @@ class Raven_Client
             $data['extra'] = array();
         }
         if (!isset($data['event_id'])) {
-            $data['event_id'] = $this->uuid4();
+            $data['event_id'] = static::uuid4();
         }
 
         if (isset($data['message'])) {
@@ -428,7 +862,7 @@ class Raven_Client
 
         $data = array_merge($this->get_default_data(), $data);
 
-        if ($this->is_http_request()) {
+        if (static::is_http_request()) {
             $data = array_merge($this->get_http_data(), $data);
         }
 
@@ -451,20 +885,28 @@ class Raven_Client
             $this->context->extra,
             $data['extra']);
 
-        $serializer = new Raven_Serializer();
-        // avoid empty arrays (which dont convert to dicts)
         if (empty($data['extra'])) {
             unset($data['extra']);
-        } else {
-            $data['extra'] = $serializer->serialize($data['extra']);
         }
         if (empty($data['tags'])) {
             unset($data['tags']);
-        } else {
-            $data['tags'] = $serializer->serialize($data['tags']);
         }
-        if (!empty($data['user'])) {
-            $data['user'] = $serializer->serialize($data['user']);
+        if (empty($data['user'])) {
+            unset($data['user']);
+        }
+        if (empty($data['request'])) {
+            unset($data['request']);
+        }
+        if (empty($data['site'])) {
+            unset($data['site']);
+        }
+
+        $existing_runtime_context = isset($data['contexts']['runtime']) ? $data['contexts']['runtime'] : array();
+        $runtime_context = array('version' => self::cleanup_php_version(), 'name' => 'php');
+        $data['contexts']['runtime'] =  array_merge($runtime_context, $existing_runtime_context);
+
+        if (!$this->breadcrumbs->is_empty()) {
+            $data['breadcrumbs'] = $this->breadcrumbs->fetch();
         }
 
         if ((!$stack && $this->auto_log_stacks) || $stack === true) {
@@ -474,19 +916,13 @@ class Raven_Client
             array_shift($stack);
         }
 
-        if (!empty($stack)) {
-            // manually trigger autoloading, as it's not done in some edge cases due to PHP bugs (see #60149)
-            if (!class_exists('Raven_Stacktrace')) {
-                spl_autoload_call('Raven_Stacktrace');
-            }
-
-            if (!isset($data['stacktrace'])) {
-                $data['stacktrace'] = array(
-                    'frames' => Raven_Stacktrace::get_stack_info(
-                        $stack, $this->trace, $this->shift_vars, $vars, $this->message_limit
-                    ),
-                );
-            }
+        if (! empty($stack) && ! isset($data['stacktrace']) && ! isset($data['exception'])) {
+            $data['stacktrace'] = array(
+                'frames' => Raven_Stacktrace::get_stack_info(
+                    $stack, $this->trace, $vars, $this->message_limit, $this->prefixes,
+                    $this->app_path, $this->excluded_app_paths, $this->serializer, $this->reprSerializer
+                ),
+            );
         }
 
         $this->sanitize($data);
@@ -495,27 +931,43 @@ class Raven_Client
         if (!$this->store_errors_for_bulk_send) {
             $this->send($data);
         } else {
-            if (empty($this->error_data)) {
-                $this->error_data = array();
-            }
-            $this->error_data[] = $data;
+            $this->_pending_events[] = $data;
         }
+
+        $this->_last_event_id = $data['event_id'];
 
         return $data['event_id'];
     }
 
     public function sanitize(&$data)
     {
-        // manually trigger autoloading, as it's not done in some edge cases due to PHP bugs (see #60149)
-        if (!class_exists('Raven_Serializer')) {
-            spl_autoload_call('Raven_Serializer');
+        // attempt to sanitize any user provided data
+        if (!empty($data['request'])) {
+            $data['request'] = $this->serializer->serialize($data['request'], 5);
+        }
+        if (!empty($data['user'])) {
+            $data['user'] = $this->serializer->serialize($data['user'], 3);
+        }
+        if (!empty($data['extra'])) {
+            $data['extra'] = $this->serializer->serialize($data['extra']);
+        }
+        if (!empty($data['tags'])) {
+            foreach ($data['tags'] as $key => $value) {
+                $data['tags'][$key] = @(string)$value;
+            }
+        }
+        if (!empty($data['contexts'])) {
+            $data['contexts'] = $this->serializer->serialize($data['contexts'], 5);
+        }
+        if (!empty($data['breadcrumbs'])) {
+            $data['breadcrumbs'] = $this->serializer->serialize($data['breadcrumbs'], 5);
         }
     }
 
     /**
      * Process data through all defined Raven_Processor sub-classes
      *
-     * @param array     $data       Associative array of data to log
+     * @param array $data Associative array of data to log
      */
     public function process(&$data)
     {
@@ -526,12 +978,10 @@ class Raven_Client
 
     public function sendUnsentErrors()
     {
-        if (!empty($this->error_data)) {
-            foreach ($this->error_data as $data) {
-                $this->send($data);
-            }
-            unset($this->error_data);
+        foreach ($this->_pending_events as $data) {
+            $this->send($data);
         }
+        $this->_pending_events = array();
         if ($this->store_errors_for_bulk_send) {
             //in case an error occurs after this is called, on shutdown, send any new errors.
             $this->store_errors_for_bulk_send = !defined('RAVEN_CLIENT_END_REACHED');
@@ -539,14 +989,44 @@ class Raven_Client
     }
 
     /**
+     * @param array $data
+     * @return string|bool
+     */
+    public function encode(&$data)
+    {
+        $message = Raven_Compat::json_encode($data);
+        if ($message === false) {
+            if (function_exists('json_last_error_msg')) {
+                $this->_lasterror = json_last_error_msg();
+            } else {
+                // @codeCoverageIgnoreStart
+                $this->_lasterror = json_last_error();
+                // @codeCoverageIgnoreEnd
+            }
+            return false;
+        }
+
+        if ($this->useCompression) {
+            $message = gzcompress($message);
+        }
+
+        // PHP's builtin curl_* function are happy without this, but the exec method requires it
+        $message = base64_encode($message);
+
+        return $message;
+    }
+
+    /**
      * Wrapper to handle encoding and sending data to the Sentry API server.
      *
      * @param array     $data       Associative array of data to log
      */
-    public function send($data)
+    public function send(&$data)
     {
-        if (is_callable($this->send_callback) && !call_user_func($this->send_callback, $data)) {
-            // if send_callback returns falsely, end native send
+        if (is_callable($this->send_callback)
+            && call_user_func_array($this->send_callback, array(&$data)) === false
+        ) {
+            // if send_callback returns false, end native send
             return;
         }
 
@@ -554,20 +1034,21 @@ class Raven_Client
             return;
         }
 
-        $message = Raven_Compat::json_encode($data);
-
-        if (function_exists("gzcompress")) {
-            $message = gzcompress($message);
+        if ($this->transport) {
+            call_user_func($this->transport, $this, $data);
+            return;
         }
-        $message = base64_encode($message); // PHP's builtin curl_* function are happy without this, but the exec method requires it
 
-        $client_string = 'raven-php/' . self::VERSION;
-        $timestamp = microtime(true);
+        // should this event be sampled?
+        if (rand(1, 100) / 100.0 > $this->sample_rate) {
+            return;
+        }
+
+        $message = $this->encode($data);
+
         $headers = array(
-            'User-Agent' => $client_string,
-            'X-Sentry-Auth' => $this->get_auth_header(
-                $timestamp, $client_string, $this->public_key,
-                $this->secret_key),
+            'User-Agent' => static::getUserAgent(),
+            'X-Sentry-Auth' => $this->getAuthHeader(),
             'Content-Type' => 'application/octet-stream'
         );
 
@@ -577,22 +1058,26 @@ class Raven_Client
     /**
      * Send data to Sentry
      *
-     * @param string    $url        Full URL to Sentry
-     * @param array     $data       Associative array of data to log
-     * @param array     $headers    Associative array of headers
+     * @param string       $url     Full URL to Sentry
+     * @param array|string $data    Associative array of data to log
+     * @param array        $headers Associative array of headers
      */
-    private function send_remote($url, $data, $headers=array())
+    protected function send_remote($url, $data, $headers = array())
     {
         $parts = parse_url($url);
         $parts['netloc'] = $parts['host'].(isset($parts['port']) ? ':'.$parts['port'] : null);
         $this->send_http($url, $data, $headers);
     }
 
-    protected function get_default_ca_cert()
+    protected static function get_default_ca_cert()
     {
         return dirname(__FILE__) . DIRECTORY_SEPARATOR . 'data' . DIRECTORY_SEPARATOR . 'cacert.pem';
     }
 
+    /**
+     * @return array
+     * @see http://stackoverflow.com/questions/9062798/php-curl-timeout-is-not-working/9063006#9063006
+     */
     protected function get_curl_options()
     {
         $options = array(
@@ -600,7 +1085,7 @@ class Raven_Client
             CURLOPT_SSL_VERIFYHOST => 2,
             CURLOPT_SSL_VERIFYPEER => $this->verify_ssl,
             CURLOPT_CAINFO => $this->ca_cert,
-            CURLOPT_USERAGENT => 'raven-php/' . self::VERSION,
+            CURLOPT_USERAGENT => 'sentry-php/' . self::VERSION,
         );
         if ($this->http_proxy) {
             $options[CURLOPT_PROXY] = $this->http_proxy;
@@ -617,7 +1102,7 @@ class Raven_Client
 
             // some versions of PHP 5.3 don't have this defined correctly
             if (!defined('CURLOPT_CONNECTTIMEOUT_MS')) {
-                //see http://stackoverflow.com/questions/9062798/php-curl-timeout-is-not-working/9063006#9063006
+                //see stackoverflow link in the phpdoc
                 define('CURLOPT_CONNECTTIMEOUT_MS', 156);
             }
 
@@ -635,11 +1120,11 @@ class Raven_Client
     /**
      * Send the message over http to the sentry url given
      *
-     * @param string $url       URL of the Sentry instance to log to
-     * @param array $data       Associative array of data to log
-     * @param array $headers    Associative array of headers
+     * @param string       $url     URL of the Sentry instance to log to
+     * @param array|string $data    Associative array of data to log
+     * @param array        $headers Associative array of headers
      */
-    private function send_http($url, $data, $headers=array())
+    protected function send_http($url, $data, $headers = array())
     {
         if ($this->curl_method == 'async') {
             $this->_curl_handler->enqueue($url, $data, $headers);
@@ -650,43 +1135,47 @@ class Raven_Client
         }
     }
 
-    /**
-     * Send the cURL to Sentry asynchronously. No errors will be returned from cURL
-     *
-     * @param string    $url        URL of the Sentry instance to log to
-     * @param array     $data       Associative array of data to log
-     * @param array     $headers    Associative array of headers
-     * @return bool
-     */
-    private function send_http_asynchronous_curl_exec($url, $data, $headers)
+    protected function buildCurlCommand($url, $data, $headers)
     {
         // TODO(dcramer): support ca_cert
         $cmd = $this->curl_path.' -X POST ';
         foreach ($headers as $key => $value) {
-            $cmd .= '-H \''. $key. ': '. $value. '\' ';
+            $cmd .= '-H ' . escapeshellarg($key.': '.$value). ' ';
         }
-        $cmd .= '-d \''. $data .'\' ';
-        $cmd .= '\''. $url .'\' ';
+        $cmd .= '-d ' . escapeshellarg($data) . ' ';
+        $cmd .= escapeshellarg($url) . ' ';
         $cmd .= '-m 5 ';  // 5 second timeout for the whole process (connect + send)
         if (!$this->verify_ssl) {
             $cmd .= '-k ';
         }
         $cmd .= '> /dev/null 2>&1 &'; // ensure exec returns immediately while curl runs in the background
 
-        exec($cmd);
+        return $cmd;
+    }
 
+    /**
+     * Send the cURL to Sentry asynchronously. No errors will be returned from cURL
+     *
+     * @param string       $url     URL of the Sentry instance to log to
+     * @param array|string $data    Associative array of data to log
+     * @param array        $headers Associative array of headers
+     * @return bool
+     */
+    protected function send_http_asynchronous_curl_exec($url, $data, $headers)
+    {
+        exec($this->buildCurlCommand($url, $data, $headers));
         return true; // The exec method is just fire and forget, so just assume it always works
     }
 
     /**
      * Send a blocking cURL to Sentry and check for errors from cURL
      *
-     * @param string    $url        URL of the Sentry instance to log to
-     * @param array     $data       Associative array of data to log
-     * @param array     $headers    Associative array of headers
+     * @param string       $url     URL of the Sentry instance to log to
+     * @param array|string $data    Associative array of data to log
+     * @param array        $headers Associative array of headers
      * @return bool
      */
-    private function send_http_synchronous($url, $data, $headers)
+    protected function send_http_synchronous($url, $data, $headers)
     {
         $new_headers = array();
         foreach ($headers as $key => $value) {
@@ -695,35 +1184,47 @@ class Raven_Client
         // XXX(dcramer): Prevent 100-continue response form server (Fixes GH-216)
         $new_headers[] = 'Expect:';
 
-        $curl = curl_init($url);
-        curl_setopt($curl, CURLOPT_POST, 1);
-        curl_setopt($curl, CURLOPT_HTTPHEADER, $new_headers);
-        curl_setopt($curl, CURLOPT_POSTFIELDS, $data);
-        curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
+        if (is_null($this->_curl_instance)) {
+            $this->_curl_instance = curl_init($url);
+        }
+        curl_setopt($this->_curl_instance, CURLOPT_POST, 1);
+        curl_setopt($this->_curl_instance, CURLOPT_HTTPHEADER, $new_headers);
+        curl_setopt($this->_curl_instance, CURLOPT_POSTFIELDS, $data);
+        curl_setopt($this->_curl_instance, CURLOPT_RETURNTRANSFER, true);
 
         $options = $this->get_curl_options();
-        $ca_cert = $options[CURLOPT_CAINFO];
-        unset($options[CURLOPT_CAINFO]);
-        curl_setopt_array($curl, $options);
-
-        curl_exec($curl);
-
-        $errno = curl_errno($curl);
-        // CURLE_SSL_CACERT || CURLE_SSL_CACERT_BADFILE
-        if ($errno == 60 || $errno == 77) {
-            curl_setopt($curl, CURLOPT_CAINFO, $ca_cert);
-            curl_exec($curl);
-        }
-
-        $code = curl_getinfo($curl, CURLINFO_HTTP_CODE);
-        $success = ($code == 200);
-        if (!$success) {
-            // It'd be nice just to raise an exception here, but it's not very PHP-like
-            $this->_lasterror = curl_error($curl);
+        if (isset($options[CURLOPT_CAINFO])) {
+            $ca_cert = $options[CURLOPT_CAINFO];
+            unset($options[CURLOPT_CAINFO]);
         } else {
-            $this->_lasterror = null;
+            $ca_cert = null;
         }
-        curl_close($curl);
+        curl_setopt_array($this->_curl_instance, $options);
+
+        $buffer = curl_exec($this->_curl_instance);
+
+        $errno = curl_errno($this->_curl_instance);
+        // CURLE_SSL_CACERT || CURLE_SSL_CACERT_BADFILE
+        if ((($errno == 60) || ($errno == 77)) && !is_null($ca_cert)) {
+            curl_setopt($this->_curl_instance, CURLOPT_CAINFO, $ca_cert);
+            $buffer = curl_exec($this->_curl_instance);
+        }
+        if ($errno != 0) {
+            $this->_lasterror = curl_error($this->_curl_instance);
+            $this->_last_sentry_error = null;
+            return false;
+        }
+
+        $code = curl_getinfo($this->_curl_instance, CURLINFO_HTTP_CODE);
+        $success = ($code == 200);
+        if ($success) {
+            $this->_lasterror = null;
+            $this->_last_sentry_error = null;
+        } else {
+            // It'd be nice just to raise an exception here, but it's not very PHP-like
+            $this->_lasterror = curl_error($this->_curl_instance);
+            $this->_last_sentry_error = @json_decode($buffer);
+        }
 
         return $success;
     }
@@ -731,13 +1232,13 @@ class Raven_Client
     /**
      * Generate a Sentry authorization header string
      *
-     * @param string    $timestamp      Timestamp when the event occurred
-     * @param string    $client         HTTP client name (not Raven_Client object)
-     * @param string    $api_key        Sentry API key
-     * @param string    $secret_key     Sentry API key
+     * @param string $timestamp  Timestamp when the event occurred
+     * @param string $client     HTTP client name (not Raven_Client object)
+     * @param string $api_key    Sentry API key
+     * @param string $secret_key Sentry API key
      * @return string
      */
-    protected function get_auth_header($timestamp, $client, $api_key, $secret_key)
+    protected static function get_auth_header($timestamp, $client, $api_key, $secret_key)
     {
         $header = array(
             sprintf('sentry_timestamp=%F', $timestamp),
@@ -757,12 +1258,20 @@ class Raven_Client
         return sprintf('Sentry %s', implode(', ', $header));
     }
 
+    public function getAuthHeader()
+    {
+        $timestamp = microtime(true);
+        return $this->get_auth_header(
+            $timestamp, static::getUserAgent(), $this->public_key, $this->secret_key
+        );
+    }
+
     /**
      * Generate an uuid4 value
      *
      * @return string
      */
-    private function uuid4()
+    protected static function uuid4()
     {
         $uuid = sprintf('%04x%04x-%04x-%04x-%04x-%04x%04x%04x',
             // 32 bits for "time_low"
@@ -804,6 +1313,13 @@ class Raven_Client
             : (!empty($_SERVER['LOCAL_ADDR'])  ? $_SERVER['LOCAL_ADDR']
             : (!empty($_SERVER['SERVER_ADDR']) ? $_SERVER['SERVER_ADDR'] : '')));
 
+        if (!$this->ignore_server_port) {
+            $hasNonDefaultPort = !empty($_SERVER['SERVER_PORT']) && !in_array((int)$_SERVER['SERVER_PORT'], array(80, 443));
+            if ($hasNonDefaultPort && !preg_match('#:[0-9]*$#', $host)) {
+                $host .= ':' . $_SERVER['SERVER_PORT'];
+            }
+        }
+
         $httpS = $this->isHttps() ? 's' : '';
         return "http{$httpS}://{$host}{$_SERVER['REQUEST_URI']}";
     }
@@ -824,8 +1340,8 @@ class Raven_Client
         }
 
         if (!empty($this->trust_x_forwarded_proto) &&
-            !empty($_SERVER['X-FORWARDED-PROTO']) &&
-            $_SERVER['X-FORWARDED-PROTO'] === 'https') {
+            !empty($_SERVER['HTTP_X_FORWARDED_PROTO']) &&
+            $_SERVER['HTTP_X_FORWARDED_PROTO'] === 'https') {
             return true;
         }
 
@@ -835,10 +1351,10 @@ class Raven_Client
     /**
      * Get the value of a key from $_SERVER
      *
-     * @param string $key       Key whose value you wish to obtain
-     * @return string           Key's value
+     * @param string $key Key whose value you wish to obtain
+     * @return string     Key's value
      */
-    private function _server_variable($key)
+    private static function _server_variable($key)
     {
         if (isset($_SERVER[$key])) {
             return $_SERVER[$key];
@@ -850,8 +1366,8 @@ class Raven_Client
     /**
      * Translate a PHP Error constant into a Sentry log level group
      *
-     * @param string $severity  PHP E_$x error constant
-     * @return string           Sentry log level group
+     * @param string $severity PHP E_$x error constant
+     * @return string          Sentry log level group
      */
     public function translateSeverity($severity)
     {
@@ -873,7 +1389,7 @@ class Raven_Client
             case E_STRICT:             return Raven_Client::INFO;
             case E_RECOVERABLE_ERROR:  return Raven_Client::ERROR;
         }
-        if (version_compare(PHP_VERSION, '5.3.0', '>=')) {
+        if (PHP_VERSION_ID >= 50300) {
             switch ($severity) {
             case E_DEPRECATED:         return Raven_Client::WARN;
             case E_USER_DEPRECATED:    return Raven_Client::WARN;
@@ -896,11 +1412,13 @@ class Raven_Client
     /**
      * Convenience function for setting a user's ID and Email
      *
-     * @param string $id            User's ID
-     * @param string|null $email    User's email
-     * @param array $data           Additional user data
+     * @deprecated
+     * @param string      $id    User's ID
+     * @param string|null $email User's email
+     * @param array       $data  Additional user data
+     * @codeCoverageIgnore
      */
-    public function set_user_data($id, $email=null, $data=array())
+    public function set_user_data($id, $email = null, $data = array())
     {
         $user = array('id' => $id);
         if (isset($email)) {
@@ -909,20 +1427,40 @@ class Raven_Client
         $this->user_context(array_merge($user, $data));
     }
 
+    public function onShutdown()
+    {
+        if (!defined('RAVEN_CLIENT_END_REACHED')) {
+            define('RAVEN_CLIENT_END_REACHED', true);
+        }
+        $this->sendUnsentErrors();
+        if ($this->curl_method == 'async') {
+            $this->_curl_handler->join();
+        }
+    }
+
     /**
      * Sets user context.
      *
-     * @param array $data   Associative array of user data
+     * @param array $data  Associative array of user data
+     * @param bool  $merge Merge existing context with new context
      */
-    public function user_context($data)
+    public function user_context($data, $merge = true)
     {
-        $this->context->user = $data;
+        if ($merge && $this->context->user !== null) {
+            // bail if data is null
+            if (!$data) {
+                return;
+            }
+            $this->context->user = array_merge($this->context->user, $data);
+        } else {
+            $this->context->user = $data;
+        }
     }
 
     /**
      * Appends tags context.
      *
-     * @param array $data   Associative array of tags
+     * @param array $data Associative array of tags
      */
     public function tags_context($data)
     {
@@ -932,7 +1470,7 @@ class Raven_Client
     /**
      * Appends additional context.
      *
-     * @param array $data   Associative array of extra data
+     * @param array $data Associative array of extra data
      */
     public function extra_context($data)
     {
@@ -945,5 +1483,80 @@ class Raven_Client
     public function setProcessors(array $processors)
     {
         $this->processors = $processors;
+    }
+
+    /**
+     * @return object|null
+     */
+    public function getLastSentryError()
+    {
+        return $this->_last_sentry_error;
+    }
+
+    /**
+     * @return bool
+     */
+    public function getShutdownFunctionHasBeenSet()
+    {
+        return $this->_shutdown_function_has_been_set;
+    }
+
+    public function close_curl_resource()
+    {
+        if (!is_null($this->_curl_instance)) {
+            curl_close($this->_curl_instance);
+            $this->_curl_instance = null;
+        }
+    }
+
+    /**
+     * @param Raven_Serializer $serializer
+     */
+    public function setSerializer(Raven_Serializer $serializer)
+    {
+        $this->serializer = $serializer;
+    }
+
+    /**
+     * @param Raven_ReprSerializer $reprSerializer
+     */
+    public function setReprSerializer(Raven_ReprSerializer $reprSerializer)
+    {
+        $this->reprSerializer = $reprSerializer;
+    }
+
+    /**
+     * Cleans up the PHP version string by extracting junk from the extra part of the version.
+     *
+     * @param string $extra
+     *
+     * @return string
+     */
+    public static function cleanup_php_version($extra = PHP_EXTRA_VERSION)
+    {
+        $version = PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION . '.' . PHP_RELEASE_VERSION;
+
+        if (!empty($extra) && preg_match('{^-(?<extra>(beta|rc)-?([0-9]+)?(-dev)?)}i', $extra, $matches) === 1) {
+            $version .= '-' . $matches['extra'];
+        }
+
+        return $version;
+    }
+
+    private function triggerAutoload()
+    {
+        // manually trigger autoloading, as it cannot be done during error handling in some edge cases due to PHP (see #60149)
+
+        if (! class_exists('Raven_Stacktrace')) {
+            spl_autoload_call('Raven_Stacktrace');
+        }
+
+        if (function_exists('mb_detect_encoding')) {
+            mb_detect_encoding('string');
+        }
+
+        if (function_exists('mb_convert_encoding')) {
+            mb_convert_encoding('string', 'UTF8');
+        }
     }
 }
