@@ -25,7 +25,7 @@ class Raven_Compat
         return php_uname('n');
     }
 
-    public static function hash_hmac($algo, $data, $key, $raw_output=false)
+    public static function hash_hmac($algo, $data, $key, $raw_output = false)
     {
         if (function_exists('hash_hmac')) {
             return hash_hmac($algo, $data, $key, $raw_output);
@@ -36,9 +36,15 @@ class Raven_Compat
 
     /**
      * Implementation from 'KC Cloyd'.
-     * See http://nl2.php.net/manual/en/function.hash-hmac.php
+     *
+     * @param string $algo       Name of selected hashing algorithm
+     * @param string $data       Message to be hashed
+     * @param string $key        Shared secret key used for generating the HMAC variant of the message digest
+     * @param bool   $raw_output Must be binary
+     * @return string
+     * @doc http://php.net/manual/en/function.hash-hmac.php
      */
-    public static function _hash_hmac($algo, $data, $key, $raw_output=false)
+    public static function _hash_hmac($algo, $data, $key, $raw_output = false)
     {
         $algo = strtolower($algo);
         $pack = 'H'.strlen($algo('test'));
@@ -66,25 +72,55 @@ class Raven_Compat
     /**
      * Note that we discard the options given to be compatible
      * with PHP < 5.3
+     *
+     * @param mixed $value
+     * @param int   $options
+     * @param int   $depth Set the maximum depth
+     * @return string
      */
-    public static function json_encode($value, $options=0)
+    public static function json_encode($value, $options = 0, $depth = 512)
     {
         if (function_exists('json_encode')) {
-            return json_encode($value);
+            if (PHP_VERSION_ID < 50300) {
+                return json_encode($value);
+            } elseif (PHP_VERSION_ID < 50500) {
+                return json_encode($value, $options);
+            } else {
+                return json_encode($value, $options, $depth);
+            }
         }
 
-        return self::_json_encode($value);
+        // @codeCoverageIgnoreStart
+        return self::_json_encode($value, $depth);
+        // @codeCoverageIgnoreEnd
+    }
+
+    /**
+     * @param mixed $value
+     * @param int   $depth Set the maximum depth
+     * @return string|false
+     */
+    public static function _json_encode($value, $depth = 513)
+    {
+        if (ini_get('xdebug.extended_info') !== false) {
+            ini_set('xdebug.max_nesting_level', 2048);
+        }
+        return self::_json_encode_lowlevel($value, $depth);
     }
 
     /**
      * Implementation taken from
      * http://www.mike-griffiths.co.uk/php-json_encode-alternative/
+     *
+     * @param mixed $value
+     * @param int   $depth Set the maximum depth
+     * @return string|false
      */
-    public static function _json_encode($value)
+    private static function _json_encode_lowlevel($value, $depth)
     {
         static $jsonReplaces = array(
-            array('\\', '/', "\n", "\t", "\r", "\b", "\f", '"'),
-            array('\\\\', '\\/', '\\n', '\\t', '\\r', '\\b', '\\f', '\"'));
+            array('\\', '/', "\n", "\t", "\r", "\f", '"'),
+            array('\\\\', '\\/', '\\n', '\\t', '\\r', '\\f', '\"'));
 
         if (is_null($value)) {
             return 'null';
@@ -97,7 +133,6 @@ class Raven_Compat
         }
 
         if (is_scalar($value)) {
-
             // Always use '.' for floats.
             if (is_float($value)) {
                 return floatval(str_replace(',', '.', strval($value)));
@@ -108,6 +143,8 @@ class Raven_Compat
             } else {
                 return $value;
             }
+        } elseif ($depth <= 1) {
+            return false;
         }
 
         $isList = true;
@@ -120,13 +157,21 @@ class Raven_Compat
         $result = array();
         if ($isList) {
             foreach ($value as $v) {
-                $result[] = self::_json_encode($v);
+                $this_value = self::_json_encode($v, $depth - 1);
+                if ($this_value === false) {
+                    return false;
+                }
+                $result[] = $this_value;
             }
 
             return '[' . join(',', $result) . ']';
         } else {
             foreach ($value as $k => $v) {
-                $result[] = self::_json_encode($k) . ':' . self::_json_encode($v);
+                $this_value = self::_json_encode($v, $depth - 1);
+                if ($this_value === false) {
+                    return false;
+                }
+                $result[] = self::_json_encode($k, $depth - 1).':'.$this_value;
             }
 
             return '{' . join(',', $result) . '}';
