@@ -17,19 +17,19 @@
 // TODO(dcramer): handle ca_cert
 class Raven_CurlHandler
 {
-    private $join_timeout;
-    private $multi_handle;
-    private $options;
-    private $requests;
+    protected $join_timeout;
+    protected $multi_handle;
+    protected $options;
+    protected $requests;
 
-    public function __construct($options, $join_timeout=5)
+    public function __construct($options, $join_timeout = 5)
     {
         $this->options = $options;
         $this->multi_handle = curl_multi_init();
         $this->requests = array();
-        $this->join_timeout = 5;
+        $this->join_timeout = $join_timeout;
 
-        register_shutdown_function(array($this, 'join'));
+        $this->registerShutdownFunction();
     }
 
     public function __destruct()
@@ -37,7 +37,7 @@ class Raven_CurlHandler
         $this->join();
     }
 
-    public function enqueue($url, $data=null, $headers=array())
+    public function enqueue($url, $data = null, $headers = array())
     {
         $ch = curl_init();
 
@@ -69,7 +69,12 @@ class Raven_CurlHandler
         return $fd;
     }
 
-    public function join($timeout=null)
+    public function registerShutdownFunction()
+    {
+        register_shutdown_function(array($this, 'join'));
+    }
+
+    public function join($timeout = null)
     {
         if (!isset($timeout)) {
             $timeout = $this->join_timeout;
@@ -84,9 +89,13 @@ class Raven_CurlHandler
         } while ($timeout !== 0 && time() - $start < $timeout);
     }
 
-    // http://se2.php.net/manual/en/function.curl-multi-exec.php
-    private function select()
+    /**
+     * @doc http://php.net/manual/en/function.curl-multi-exec.php
+     */
+    protected function select()
     {
+        $active = false;
+        
         do {
             $mrc = curl_multi_exec($this->multi_handle, $active);
         } while ($mrc == CURLM_CALL_MULTI_PERFORM);
