@@ -10,7 +10,7 @@
  */
 
 /**
- * Serializes a value into a representation that should reasonable suggest
+ * Serializes a value into a representation that should reasonably suggest
  * both the type and value, and be serializable into JSON.
  * @package raven
  */
@@ -35,13 +35,7 @@ class Raven_ReprSerializer extends Raven_Serializer
         } elseif (is_array($value)) {
             return 'Array of length ' . count($value);
         } else {
-            $value = (string) $value;
-
-            if (function_exists('mb_convert_encoding')) {
-                $value = mb_convert_encoding($value, 'UTF-8', 'auto');
-            }
-
-            return $value;
+            return $this->serializeString($value);
         }
     }
 }
