@@ -26,26 +26,100 @@
  */
 class Raven_Serializer
 {
+    /*
+     * The default mb detect order
+     *
+     * @see http://php.net/manual/en/function.mb-detect-encoding.php
+     */
+    const DEFAULT_MB_DETECT_ORDER = 'auto';
+
+    /*
+     * Suggested detect order for western countries
+     */
+    const WESTERN_MB_DETECT_ORDER = 'UTF-8, ASCII, ISO-8859-1, ISO-8859-2, ISO-8859-3, ISO-8859-4, ISO-8859-5, ISO-8859-6, ISO-8859-7, ISO-8859-8, ISO-8859-9, ISO-8859-10, ISO-8859-13, ISO-8859-14, ISO-8859-15, ISO-8859-16, Windows-1251, Windows-1252, Windows-1254';
+
+    /**
+     * This is the default mb detect order for the detection of encoding
+     *
+     * @var string
+     */
+    protected $mb_detect_order = self::DEFAULT_MB_DETECT_ORDER;
+
+    /**
+     * The default maximum message lengths. Longer strings will be truncated
+     *
+     * @var int
+     */
+    protected $message_limit = Raven_Client::MESSAGE_LIMIT;
+
+    /**
+     * @param null|string $mb_detect_order
+     * @param null|int    $message_limit
+     */
+    public function __construct($mb_detect_order = null, $message_limit = null)
+    {
+        if ($mb_detect_order != null) {
+            $this->mb_detect_order = $mb_detect_order;
+        }
+
+        if ($message_limit != null) {
+            $this->message_limit = (int) $message_limit;
+        }
+    }
+
     /**
      * Serialize an object (recursively) into something safe for data
      * sanitization and encoding.
+     *
+     * @param mixed $value
+     * @param int   $max_depth
+     * @param int   $_depth
+     * @return string|bool|double|int|null|object|array
      */
-    public function serialize($value, $max_depth=3, $_depth=0)
+    public function serialize($value, $max_depth = 3, $_depth = 0)
     {
-        if (is_object($value) || is_resource($value)) {
-            return $this->serializeValue($value);
-        } elseif ($_depth < $max_depth && is_array($value)) {
+        $className = is_object($value) ? get_class($value) : null;
+        $toArray = is_array($value) || $className === 'stdClass';
+        if ($toArray && $_depth < $max_depth) {
             $new = array();
             foreach ($value as $k => $v) {
                 $new[$this->serializeValue($k)] = $this->serialize($v, $max_depth, $_depth + 1);
             }
 
             return $new;
-        } else {
-            return $this->serializeValue($value);
         }
+        return $this->serializeValue($value);
     }
 
+    protected function serializeString($value)
+    {
+        $value = (string) $value;
+
+        // Check if mbstring extension is loaded
+        if (extension_loaded('mbstring')) {
+            // we always guarantee this is coerced, even if we can't detect encoding
+            if ($currentEncoding = mb_detect_encoding($value, $this->mb_detect_order)) {
+                $value = mb_convert_encoding($value, 'UTF-8', $currentEncoding);
+            } else {
+                $value = mb_convert_encoding($value, 'UTF-8');
+            }
+
+            if ($this->message_limit !== 0 && mb_strlen($value) > $this->message_limit) {
+                $value = mb_substr($value, 0, $this->message_limit - 10, 'UTF-8') . ' {clipped}';
+            }
+        } else {
+            if ($this->message_limit !== 0 && strlen($value) > $this->message_limit) {
+                $value = substr($value, 0, $this->message_limit - 10) . ' {clipped}';
+            }
+        }
+
+        return $value;
+    }
+
+    /**
+     * @param mixed $value
+     * @return string|bool|double|int|null
+     */
     protected function serializeValue($value)
     {
         if (is_null($value) || is_bool($value) || is_float($value) || is_integer($value)) {
@@ -57,13 +131,48 @@ class Raven_Serializer
         } elseif (is_array($value)) {
             return 'Array of length ' . count($value);
         } else {
-            $value = (string) $value;
-
-            if (function_exists('mb_convert_encoding')) {
-                $value = mb_convert_encoding($value, 'UTF-8', 'auto');
-            }
-
-            return $value;
+            return $this->serializeString($value);
         }
+    }
+
+
+    /**
+     * @return string
+     * @codeCoverageIgnore
+     */
+    public function getMbDetectOrder()
+    {
+        return $this->mb_detect_order;
+    }
+
+    /**
+     * @param string $mb_detect_order
+     *
+     * @return Raven_Serializer
+     * @codeCoverageIgnore
+     */
+    public function setMbDetectOrder($mb_detect_order)
+    {
+        $this->mb_detect_order = $mb_detect_order;
+
+        return $this;
+    }
+
+    /**
+     * @return int
+     * @codeCoverageIgnore
+     */
+    public function getMessageLimit()
+    {
+        return $this->message_limit;
+    }
+
+    /**
+     * @param int $message_limit
+     * @codeCoverageIgnore
+     */
+    public function setMessageLimit($message_limit)
+    {
+        $this->message_limit = (int)$message_limit;
     }
 }
