@@ -21,8 +21,13 @@ class Raven_Util
      * Because we love Python, this works much like dict.get() in Python.
      *
      * Returns $var from $array if set, otherwise returns $default.
+     *
+     * @param array  $array
+     * @param string $var
+     * @param mixed  $default
+     * @return mixed
      */
-    public static function get($array, $var, $default=null)
+    public static function get($array, $var, $default = null)
     {
         if (isset($array[$var])) {
             return $array[$var];
